# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend for a video-based social commerce platform.\n\n## Authentication\nThis API uses **JWT Bearer** authentication. Obtain a token via\n`POST /api/v1/auth/register` or `POST /api/v1/auth/login`, then pass it\nin the `Authorization` header:\n```\nAuthorization: Bearer \u003ctoken\u003e\n```\n\n## Roles\n| Role  | Description |\n|-------|-------------|\n| user  | Default role assigned on registration |\n| admin | Required for admin-only endpoints (e.g. `GET /api/v1/admin/users`) |\n\n## Optional Authentication\nSome endpoints (`GET /feed`, `GET /videos/:id`, `POST /videos/:id/view`,\n`GET /videos/:id/stats`) accept but do not require a JWT. Anonymous\nrequests are served normally; authenticated requests may receive\npersonalised responses in future versions.",
        "title": "Video Commerce \u0026 Social Feed API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/feed": {
            "get": {
                "description": "Returns a cursor-paginated list of videos ordered by creation time (newest first).\nAuthentication is optional — when a valid JWT is supplied the caller is identified\nfor future personalisation features.\n\n**Cursor pagination:** pass the `next_cursor` value from the previous response as\nthe `cursor` query parameter to fetch the next page. An empty `next_cursor` in the\nresponse means there are no more pages.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feed"
                ],
                "summary": "Get paginated video feed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token} — optional",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 timestamp cursor from previous page",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (1–100, default 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.FeedResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid cursor format",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/videos": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new video/reel owned by the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Videos"
                ],
                "summary": "Upload a video",
                "parameters": [
                    {
                        "description": "Video metadata",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateVideoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entities.Video"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid JWT",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/videos/{id}": {
            "get": {
                "description": "Returns a single video. Accessible by anyone. If a valid JWT is provided the request is attributed to that user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Videos"
                ],
                "summary": "Get a video by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token} — optional, enables per-user features",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entities.Video"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Video not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/videos/{id}/like": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds a like from the authenticated user to the specified video.\nReturns 409 if the video is already liked by this user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interactions"
                ],
                "summary": "Like a video",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entities.Like"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Missing or invalid JWT",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Video not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "409": {
                        "description": "Video already liked by this user",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/videos/{id}/stats": {
            "get": {
                "description": "Returns view count, like count, and engagement rate for the specified video.\nEngagement rate = (likes / views) × 100. Returns 0 when views = 0.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get video analytics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token} — optional",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/videousecase.VideoStats"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Video not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/videos/{id}/unlike": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes the authenticated user's like from the specified video.\nReturns 404 if the user has not liked this video.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interactions"
                ],
                "summary": "Unlike a video",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video unliked successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid JWT",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Like not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/videos/{id}/view": {
            "post": {
                "description": "Records a view for the specified video. Authentication is optional.\n\n**Deduplication window: 1 hour.**\n- Authenticated: deduplicated per `userID + videoID`.\n- Anonymous: deduplicated per `clientIP + videoID`.\n\nThe response field `counted` indicates whether this call actually\nincremented the view counter (false = duplicate within the window).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interactions"
                ],
                "summary": "Track a video view",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token} — optional",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.TrackViewResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Video not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.Like": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "post_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "entities.Video": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "likes_count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                },
                "views_count": {
                    "type": "integer"
                }
            }
        },
...
        "request.CreateVideoRequest": {
            "type": "object",
            "required": [
                "title",
                "video_url"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
...
        "response.FeedResponseData": {
            "type": "object",
            "properties": {
                "nextCursor": {
                    "type": "string"
                },
                "videos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Video"
                    }
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.TrackViewResponseData": {
            "type": "object",
            "properties": {
                "counted": {
                    "type": "boolean"
                }
            }
        },
...
        "videousecase.VideoStats": {
            "type": "object",
            "properties": {
                "engagement_rate": {
                    "type": "number"
                },
                "likes": {
                    "type": "integer"
                },
                "video_id": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT as: **Bearer \u0026lt;token\u0026gt;**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}